feat: rewrite trashcan buttons to be persistent between bot restarts
This solves problems with deletion buttons having race conditions or
being replaced by a newer view. This change also encodes the allowed
user in the custom_id along with an override of allowed permissions.

Futhermore, the code has been simplifed quite a lot. There is now only a
need to pass the allowed user to the view on creation. Once this is set,
the button, when pressed, will be handled by a specific listener.