feat: implement bot features -- a new subsystem
this is a new system which allows turning on and off components at will
these are configurable either globally or per-guild
this allows for more precision and the ability to enable or disable a
rouge component at any time

This also allows more bot functions which are in beta but not ready to
be exposed to all guilds, but are still useful. Some of the features
here may never be made globally available, but this system allows that
possibilty, as its entire point.