chore: squash black & isort into ruff (#647)
This project uses `ruff`, `black` and `isort`. However, ruff can handle
all the logic for everything. And so this PR squashes them all into one
thing.

With this change, I also update the target python style version. The
only meaningful change there is adding strict to zip. Let me know if you
want these zip functions to be strict or not.

One note is that the bot can also be updated to format with ruff instead
of black. But that is a larger feature change and not in the scope of
this PR.

---------

Co-authored-by: arielle <me@arielle.codes>