fix(codesnippets): use `?per_page=100` when fetching branches/tags (#682)
Auto-linking of code snippets fetches branches and tags in an attempt to
partition the `blob/things/stuff/path/file.py` string into a known
branch and file path, but previously only fetched the first 30, which is
the default.
This changes it to 100, the max before we'd have to start paginating.
It's not perfect, but at least more reliable than before c: