refactor: use the pep API for gathering PEP information (#685)
Instead of using the sphinx object inventory, we can get all of the peps
from a single generated API, which includes all of the information we
used to get from sphinx. This also saves on requesting PEP information
to get the data, as all of the data we consumed is hosted in the API.

This leaves fetching and parsing the contents of the PEP still in place,
so information can be gathered about the headings on a PEP itself.